Defined some behavior.
Commented some methods in Server class.
Added some methods to Database class.
Some thoughts: Instead of using a wallet, why not just use a member
variable to represent money not in the bank that player has on hand?
Instead of using singletons, why not use static classes? Instead of
having user names and user IDs, why not use one or the other? Why not
use unsigned ints to represent all amounts in cents, and convert to
dollars and cents when necessary? Why is the authentication method not a
Server method?