Just comments.
Notes to self, please ignore: Haven't found open source/free to use/
free of charge C# SQLite security or encryption API. Maybe
encrypt/decrypt data when written to/read from database? What's the best
way to do that? Would all of the columns just be BLOBS so that bytes
that don't correspond to printable characters can be stored? Should the
column name be unhelpful or misleading? Also, how do we stop an attacker
from opening the file and messing with the raw data?